In C++ float is too small

+++ src/loss_function.h
@@ -6,9 +6,9 @@
 
 class LossFunction {
  public:
-  LossFunction(const float l1, const float l2) {}
-  const float get_decision(const Item& item, const Weights& weights) {return 0.;}
-  const float get_loss(const Item& item, const Weights& weights) {return 0;}
+  LossFunction(const double l1, const double l2) {}
+  const double get_decision(const Item& item, const Weights& weights) {return 0.;}
+  const double get_loss(const Item& item, const Weights& weights) {return 0;}
   // Do we need it? Do we need logistic loss internally?
   const Weights get_grad(const Item& item, const Weights& weight) {return Weights();}
 };
